Refactor Gemini API handling and move logic to dedicated package
- Moved Gemini API URL building and response handling to `api/llm/gemini` package.
- Renamed and updated test file to match new package structure.
- Removed duplicate code and standardized error messages.
- Added `BuildAPIURL` function to handle URL construction with environment variable expansion.
